requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
instructor>=1.6.0
google-genai>=1.0.0
pydantic>=2.9.0
python-dotenv>=1.0.1
email-validator>=2.2.0
instructor[google-genai]>=1.6.0
